pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "market-data-collector"
version = "0.1.0"
description = "Polygon market data collector"
authors = [{name="ACME"}]
requires-python = ">=3.8"
readme = "README.md"
dependencies = [
    "requests",
    "websocket-client",
    "aiohttp",
    "pandas",
    "numpy",
    "scikit-learn",
    "lightgbm",
    "arch",
    "pytest-asyncio",
    "websockets",
    "plotly",
]

[project.scripts]
collect-data = "trading_platform.collector.main:main"

[tool.setuptools.packages.find]
where = ["src", ""]
